+++ radio@hslbck.gmail.com/extension.js
@@ -8,11 +8,7 @@ const Extension = imports.misc.extensionUtils.getCurrentExtension();
 const Convenience = Extension.imports.convenience;
 const RadioMenu = Extension.imports.radioMenu;
 
-// ToDo: move to radio menu for settings
-const RadioSearchProvider = Extension.imports.searchProvider;
-
 var radioMenu;
-var radioSearchProvider;
 
 // init with translation support
 function init() {
@@ -23,22 +19,14 @@ function init() {
 function enable() {
     radioMenu = new RadioMenu.RadioMenuButton();
     Main.panel.addToStatusArea('radioMenu', radioMenu);
-
-    if (!radioSearchProvider) {
-        radioSearchProvider = new RadioSearchProvider.RadioSearchProvider();
-        Main.overview.viewSelector._searchResults._registerProvider(radioSearchProvider);
-    }
 }
 
 //  stop playing and destroy extension content
 function disable() {
     // stop playing before destruction
     // affects lock screen
-    if (radioSearchProvider) {
-        Main.overview.viewSelector._searchResults._unregisterProvider(radioSearchProvider);
-        radioSearchProvider = null;
-    }
     radioMenu._stop();
     radioMenu._disconnectMediaKeys();
+    radioMenu._disableSearchProvider();
     radioMenu.destroy();
 }